requirements.txt
# CLI框架
click>=8.1.0
rich>=13.0.0
typer>=0.9.0

# AI集成
anthropic>=0.18.0
openai>=1.12.0
requests>=2.31.0

# 文件处理
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-magic-bin>=0.4.14; sys_platform == 'win32'
python-magic>=0.4.27; sys_platform != 'win32'
Pillow>=10.0.0

# 工具库
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
